+++ cookbooks/rightimage/attributes/default.rb
@@ -91,7 +91,7 @@ end
 guest_packages =
 case rightimage[:platform]
 when "ubuntu" then %w(acpid openssh-client openssh-server language-selector-common ntp ubuntu-standard rightimage-extras-base)
-when "centos", "rhel" then %w(acpid ntp openssh-server openssl dhclient rightimage-extras-base)
+when "centos", "rhel" then %w(acpid ntp openssh-clients openssh-server openssl psmisc dhclient rightimage-extras-base)
 end
 set[:rightimage][:guest_packages] = guest_packages