+++ cookbooks/rightimage/providers/cloud_google.rb
@@ -4,11 +4,6 @@ end
 
 action :configure do
 
-  node.override[:rightimage][:grub][:root_device] = "/dev/sda"
-  node.override[:rightimage][:grub][:kernel][:options] = "noquiet earlyprintk=ttyS0 loglevel=8"
-  node.override[:rightimage][:root_mount][:dump] = "1"
-  node.override[:rightimage][:root_mount][:fsck] = "1"
-
   ruby_block "check hypervisor" do
     block do
       raise "ERROR: you must set your hypervisor to kvm!" unless new_resource.hypervisor == "kvm"
@@ -20,43 +15,58 @@ action :configure do
     command "chroot #{guest_root} yum -y install iscsi-initiator-utils"
   end
 
-  if (new_resource.platform =~ /centos|rhel/ && new_resource.platform_version.to_f >= 6) || new_resource.platform == "ubuntu"
-    # Add google init script for centos (6+ only) / ubuntu
-    cookbook_file "#{guest_root}/etc/init/google.conf" do
-      source "google.conf"
-      owner "root"
-      group "root"
-      mode "0755"
-      action :create
+  # See https://github.com/GoogleCloudPlatform/compute-image-packages for google startup
+  # scripts for all the various platforms. Prepared the tarball with:
+  #   git clone https://github.com/GoogleCloudPlatform/compute-image-packages
+  #   cd compute-image-packages/google-startup-scripts
+  #   tar --exclude etc/rc.local -zcvf ../google-startup-scripts-master-20140718.tar.gz etc lib usr
+  if (new_resource.platform =~ /centos|rhel/ && new_resource.platform_version.to_f >= 5) || new_resource.platform == "ubuntu"
+    # Add google init scripts. All system startup types (upstart, systemd, sysvinit)
+    # get extracted, though they will only be used as needed. We still need to 
+    remote_file "/tmp/google-startup-scripts.tar.gz" do
+      source "#{node[:rightimage][:s3_base_url]}/files/google-startup-scripts-v1.1.6.tar.gz"
+      action :create_if_missing
       backup false
     end
-    cookbook_file "#{guest_root}/etc/init/google_run_startup_scripts.conf" do
-      source "google_run_startup_scripts.conf"
-      owner "root"
-      group "root"
-      mode "0755"
-      action :create
-      backup false
+
+    bash "extract google startup scripts" do
+      flags "-ex"
+      code <<-EOF
+        if [ -x #{guest_root}/bin/systemctl ]; then
+          echo "Installing systemd google startup scripts"
+          exclusions="--exclude etc/init.d --exclude etc/init"
+        elif [ -x #{guest_root}/sbin/initctl ]; then 
+          echo "Installing upstart google startup scripts"
+          exclusions="--exclude etc/init.d --exclude usr/lib/systemd"
+        else
+          echo "Installing sysvinit google startup scripts"
+          exclusions="--exclude etc/init --exclude usr/lib/systemd"
+        fi
+        tar $exclusions -C #{guest_root}/ -zhxvf /tmp/google-startup-scripts.tar.gz
+      EOF
     end
-    # implement support for disk path aliases (w-5221)
-    cookbook_file "#{guest_root}/lib/udev/rules.d/65-gce-disk-naming.rules" do
-      source "google_disk_naming_rules"
-      owner "root"
-      group "root"
-      backup false
+
+    # Add the needed upstart job emission to /etc/rc.local, conditional on upstart
+    # being present, but only if it's not already in the file. stolen from the .rpm
+    # postinstall scriptlet
+    # Note that this comes after and appends the /etc/rc.local written in KVM provider
+    bash "finish google startup scripts installation" do
+      flags "-ex"
+      code <<-EOF
+        if [ -x #{guest_root}/bin/systemctl ]; then
+          chroot #{guest_root} systemctl enable google-startup-scripts.service
+          chroot #{guest_root} systemctl enable google.service
+        elif [ -x #{guest_root}/sbin/initctl ]; then 
+          grep -q 'google-rc-local-has-run' #{guest_root}/etc/rc.local && exit 0
+          echo "initctl emit --no-wait google-rc-local-has-run" >> #{guest_root}/etc/rc.local
+          chmod 755 #{guest_root}/etc/rc.local
+        fi
+      EOF
     end
   else
     raise "Unsupported platform/version combination #{new_resource.platform} #{new_resource.platform_version}"
   end
 
-  cookbook_file "#{target_raw_root}/google.tgz" do
-    source "google.tgz"
-    action :create
-    backup false
-  end
-  bash "untar google helper and startup scripts" do
-    code "tar zxf #{target_raw_root}/google.tgz -C #{guest_root}/usr/share"
-  end
 
   bash "configure for google compute" do
     flags "-ex" 
@@ -80,7 +90,9 @@ action :configure do
         # NOTE: this image should not be used in production!!
         # Precise kernel doesn't support SCSI_VIRTIO driver.
         # Quantal kernel doesn't show attached volumes. (w-6223)
-        chroot $guest_root apt-get -y install linux-generic-lts-raring
+        if [ "#{node[:rightimage][:platform_version]}" == "12.04" ]; then
+          chroot $guest_root apt-get -y install linux-generic-lts-raring
+        fi
 
         # Disable all ttys except for tty1 (console)
         for i in `ls $guest_root/etc/init/tty[2-9].conf`; do
@@ -89,16 +101,9 @@ action :configure do
         ;;
       esac
 
-      # Emit signal to run google_run_startup_scripts
-      # Note that this comes after and replaces the /etc/rc.local written in KVM provider
-      # will not work centos 5
-      echo '#!/bin/bash' > $guest_root/etc/rc.local
-      echo 'initctl emit --no-wait google-rc-local-has-run' >> $guest_root/etc/rc.local
-      chmod 755 $guest_root/etc/rc.local
-
       set +e
       # Add metadata alias
-      grep -E 'metadata' /etc/hosts &> /dev/null
+      grep -E 'metadata' $guest_root/etc/hosts &> /dev/null
       if [ "$?" != "0" ]; then
         echo '169.254.169.254 metadata.google.internal metadata' >> $guest_root/etc/hosts
       fi
@@ -184,6 +189,7 @@ action :upload do
         "--project=#{node[:rightimage][:google][:project_id]}"
       Chef::Log.info("Running command: #{command}")
       `#{command}`
+      raise "Registration failed" unless $?.success?
     end
   end