+++ cookbooks/rightimage/CHANGELOG.md
@@ -1,7 +1,8 @@
 Changes
 -------
- * 2014-06-24: bump to 14.0.14. remove floppy drive mount on vSphere, metadata is now passed via vmware tools daemon
+ * 2014-07-15: bump to 14.1.0. Chef 11 compatibility fixes. Support for Ubuntu 14.04, CentOS 7. Updated Google startup scripts to v1.1.6. Pin collectd to v4.
  * 2014-07-22: bump to 14.0.13. Point release for 14.0.1 image, vSphere Ubuntu volume support.
+ * 2014-06-24: bump to 14.0.14. remove floppy drive mount on vSphere, metadata is now passed via vmware tools daemon
  * 2014-04-24: bump to 14.0.8. update gcutil tool versions
  * 2013-10-24: added GCE Advanced OS Support for Ubuntu 12.04 with backported 12.10 kernel (beta support only!)
  * 2013-10-24: added GCE Advanced OS Support for EL6